requirements.txt
fastapi==0.115.12
supabase==2.15.1
uvicorn==0.34.2
python-dotenv==1.1.0
